fix
Tools::convertPrice expects a float ad first parameter.
as you suggested, the first (float) cast solves the problem. the second one, the one inside the if sttement, uses $price_converted_to_default_currency, wich comes from the previous line as the output of another convertPrice. The input of that convertPrice is, again not casted version of $supply_order_detail->unit_price_te
With this changes we can suppouse that when reaching second assignment of $price there will be a correct value as we can rely on convertPrice to output a safe value if the input is float.